added verifyToken and verifyAdmin middlewares and cookie with user data
