+++ src/controllers/controllers.views/dashboardUser.controller.js
@@ -1,6 +1,5 @@
 const getDashboardUser = (req,res)=>{
     //validacion de user autenticado y rol admin
-    
     res.render('homeProfileUser');
 }
