+++ src/routes/views.admin.routes.js
@@ -2,6 +2,7 @@ const express = require('express');
 const router = express.Router();
 const controllerBasePath = "../controllers/controllers.views"
 const verifyToken = require('../middlewares/verifyToken');
+const verifyAdmin = require('../middlewares/verifyAdmin');
 
 //Hacer destructuring para acotar estas líneas
 //Variables con ubicación de los controllers de views:
@@ -15,10 +16,10 @@ const editMovie = require(controllerBasePath + "/editMovie.controller");
 // Falta dashboard de admin: router.get('', dashboardAdmin)
 // router.get('/recoverpassword', login)
 // router.get('/restorepassword', login)
-router.get('/dashboardAdmin', verifyToken, dashboardAdmin.getDashboardAdmin)
-router.get('/createmovie', verifyToken, createMovie.createMovieAdmin)
+router.get('/dashboardAdmin', verifyToken, verifyAdmin, dashboardAdmin.getDashboardAdmin)
+router.get('/createmovie', verifyToken, verifyAdmin, createMovie.createMovieAdmin)
 // router.delete('/removeMovie', removeMovie)
-router.get('/editmovie', verifyToken, editMovie.editMovie);
+router.get('/editmovie', verifyToken, verifyAdmin, editMovie.editMovie);
 
 //Pendiente el post de logout (router.post('/logout', signup));
